requirements.txt

Django==4.2.7
djangorestframework==3.14.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.9.1
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.2.1
Pillow==10.1.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
django-oauth-toolkit==2.2.0
django-money==3.0.0
psycopg2-binary==2.9.9
boto3==1.24.89
django-storages==1.13.1
django-s3direct==2.0.3
python-dotenv==0.21.0
django-cors-headers==3.13.0
django-environ==0.11.2